+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.2.5.3
 
 -   Print out java version information earlier in bootup process
+-   Add in dialog panel when using Newer Java than 8